timetable-main/bot/requirements.txt
telebot~=0.0.4
pyyaml~=5.4
